Consolidated interface files. Updated root component naming to match classic React and Next.js conventions
