+++ app/src/components/bottom/CodePreview.tsx
@@ -5,7 +5,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import 'ace-builds/src-noconflict/mode-javascript';
 import 'ace-builds/src-noconflict/theme-monokai';
 import Button from '@material-ui/core/Button';
-import { Component } from '../../interfaces/InterfacesNew';
+import { Component } from '../../interfaces/Interfaces';
 
 const CodePreview = () => {
   const [state, dispatch] = useContext(stateContext);
