+++ app/src/constants/ItemTypes.ts
@@ -1,4 +1,4 @@
-import { DragItemType } from '../interfaces/interfacesNew';
+import { DragItemType } from '../interfaces/Interfaces';
 
 const ItemTypes: DragItemType = {
   INSTANCE: 'instance'
