+++ app/src/context/context.ts
@@ -1,4 +1,4 @@
 import React from 'react';
-import { State } from '../interfaces/interfacesNew';
+import { State } from '../interfaces/Interfaces';
 
 export const stateContext = React.createContext<State | {}>([{}, () => {}]);
