+++ app/src/interfaces/global.ts
@@ -1,11 +1,13 @@
 export {};
 
+// The following types are applied globally across the project
 // The 'api' key on the global window object is where preload.js
 // is storing main.js functions so they can be available to the renderer process
+
 declare global {
   interface Window {
     api: any;
   }
 }
 
-let api = window.api; 
+let api = window.api;
