+++ app/src/utils/createNextFiles.util.ts
@@ -2,7 +2,7 @@
 // "Root" level components are stored in a pages directory
 // all other components will be in a components directory
 
-import { Component } from '../interfaces/InterfacesNew';
+import { Component } from '../interfaces/Interfaces';
 
 const isRoot = (component: Component, rootArray: number[]) => {
   return rootArray.includes(component.id) ? true : false;
